requirements.txt
streamlit
langchain
python-dotenv
openai
# huggingface_hub
pypdf2
faiss-cpu
langchain-community
# sentence_transformers
# InstructorEmbedding
tiktoken